Remove GStreamer dependency and port to Qt 6
Replace the GStreamer media backend with Qt 6 Multimedia (QMediaPlayer,
QVideoSink, QAudioOutput, QCamera, QMediaCaptureSession). This removes
a heavy native dependency and brings the codebase onto a supported Qt
version.

Key changes:
- New VideoPlayerImpl class using QMediaPlayer/QVideoSink for file playback
- Rewrite CameraSurface (QVideoSink) and CameraImpl (QMediaCaptureSession)
- Drop VideoShmSrcImpl, VideoUriDecodeBinImpl, VideoV4l2SrcImpl
- QGLWidget → QOpenGLWidget, QDesktopWidget → QScreen
- QRegExp → QRegularExpression throughout (including qtpropertybrowser)
- Fix all Qt 5→6 API breakages (endl, toAscii, qCopy, QMutexLocker
  template, enterEvent, QWeakPointer::data, QVariant helpers, etc.)
- Update build scripts for Qt 6 on macOS

+++ src/control/qosc/oscreceiver.cpp
@@ -26,7 +26,7 @@ void OscReceiver::readyReadCb() {
 }
 
 void OscReceiver::byteArrayToVariantList(QVariantList& outputVariantList, QString& outputOscAddress, const QByteArray& inputByteArray) {
-    osc::ReceivedPacket packet(inputByteArray.data(), inputByteArray.size());
+    osc::ReceivedPacket packet(inputByteArray.data(), static_cast<std::size_t>(inputByteArray.size()));
     // TODO: catch parsing exceptions
     if (packet.IsMessage()) {
         osc::ReceivedMessage message(packet);
